+++ android/src/main/kotlin/be/tramckrijte/workmanager/WorkmanagerPlugin.kt
@@ -1,29 +1,41 @@
 package be.tramckrijte.workmanager
 
+import android.content.Context
 import io.flutter.embedding.android.FlutterEngineConfigurator
 import io.flutter.embedding.engine.plugins.FlutterPlugin
+import io.flutter.plugin.common.BinaryMessenger
 import io.flutter.plugin.common.MethodCall
 import io.flutter.plugin.common.MethodChannel
 import io.flutter.plugin.common.MethodChannel.MethodCallHandler
 import io.flutter.plugin.common.MethodChannel.Result
+import io.flutter.plugin.common.PluginRegistry
 
 class WorkmanagerPlugin : MethodCallHandler, FlutterPlugin {
 
     private lateinit var workmanagerCallHandler: WorkmanagerCallHandler
 
-    override fun onAttachedToEngine(binding: FlutterPlugin.FlutterPluginBinding) {
-        val channel = MethodChannel(binding.binaryMessenger, "be.tramckrijte.workmanager/foreground_channel_work_manager")
-        channel.setMethodCallHandler(WorkmanagerPlugin().apply { workmanagerCallHandler = WorkmanagerCallHandler(binding.applicationContext) })
-    }
+    override fun onAttachedToEngine(binding: FlutterPlugin.FlutterPluginBinding) =
+            registerWorkManager(binding.binaryMessenger, binding.applicationContext)
 
     override fun onDetachedFromEngine(binding: FlutterPlugin.FlutterPluginBinding) {}
 
     companion object {
-        lateinit var engineConfigurator: FlutterEngineConfigurator
+        var pluginRegistryCallback: PluginRegistry.PluginRegistrantCallback? = null
 
         @JvmStatic
-        fun setPluginRegistrantCallback(engineConfigurator: FlutterEngineConfigurator) {
-            WorkmanagerPlugin.engineConfigurator = engineConfigurator
+        private fun registerWorkManager(messenger: BinaryMessenger, ctx: Context) {
+            val channel = MethodChannel(messenger, "be.tramckrijte.workmanager/foreground_channel_work_manager")
+            channel.setMethodCallHandler(WorkmanagerPlugin().apply { workmanagerCallHandler = WorkmanagerCallHandler(ctx) })
+        }
+
+        @JvmStatic
+        fun registerWith(registrar: PluginRegistry.Registrar) =
+                registerWorkManager(registrar.messenger(), registrar.activeContext())
+
+        @Deprecated(message = "Use the Android v2 embedding method.")
+        @JvmStatic
+        fun setPluginRegistrantCallback(pluginRegistryCallback: PluginRegistry.PluginRegistrantCallback) {
+            WorkmanagerPlugin.pluginRegistryCallback = pluginRegistryCallback
         }
     }