+++ example/android/app/src/main/kotlin/be/tramckrijte/workmanager_example/EmbeddingV1Application.kt
@@ -1,0 +1,18 @@
+package be.tramckrijte.workmanager_example
+
+import be.tramckrijte.workmanager.WorkmanagerPlugin
+import io.flutter.app.FlutterApplication
+import io.flutter.plugin.common.PluginRegistry
+import io.flutter.plugins.GeneratedPluginRegistrant
+
+@Deprecated(message = "Not used, but here to show you how you can use the plugin using the old v1 embedding method.")
+class EmbeddingV1Application : FlutterApplication(), PluginRegistry.PluginRegistrantCallback {
+    override fun registerWith(registry: PluginRegistry) {
+        GeneratedPluginRegistrant.registerWith(registry)
+    }
+
+    override fun onCreate() {
+        super.onCreate()
+        WorkmanagerPlugin.setPluginRegistrantCallback(this)
+    }
+}