+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # 0.2.0
 * Android:
-    * Dropping support for registering plugins via the `registerWith` way instead only support the `Android v2 embedding`.  
+    * Adding support for the `Android v2 embedding`.  
       See more details in the Android setup guide.
     
 # 0.1.5